Move records test to tests package
* Move the tests
* Separate the tests from the Record class to test
